BOSA-7: revert the rich editor for create a new initiative
After enabling this feature I have found the real reason why it was
disabled. The data between step are stored in the session. In cases long
text with many html entities we are receiving the `ActionDispatch::Cookies::CookieOverflow`.

In this case I'm reverting to the previous state, since the fix is not a
trivial one.