ops: BOSA21Q1-230: bosa-assets doesn't serve the whole public dir
This is  a followup of 05e31696e7681bf59a328bdcfe4a04f7b02203f6

Accessing /robots.txt in uat doesn't trigger an internal server error
anymore, but a 404. The expected result was to display the robots.txt
file. By checking the nginx log, we know that this file was actually not
found on the server. By connecting on the server, we saw the only the
file was missing on the server, only the subdirectory "assets" was
found.

It due to how the image is built, where we made available only the
assets directory and not the whole public one.