BOSA21Q1-492: tests: script fetching tests from upstream
One of the possibility to improve our test coverage is to take back the
tests from decidim and adapt them since we customized so much of it. We
add some tooling in order to ease that process. The tooling will keep
track of which commits are applied and will allow to automate fetching
newer ones when moving from one decidim release to another.

The script added in this commit will loop over the enabled modules. If
it is the first time it runs for a module, it will copy paste the
content of the spec folder and write the revision in "decidim_rev.txt".
If it is ran again once the folder is already pasted, it will manually
apply the revisions touching these files, stop for a conflict if needed,
and lastly commit them. The last applied revision will again be in
"decidim_rev.txt".